backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0
